Adds merged projects windsor and aspnet to the solution
