--- NuGet.config
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <packageSources>
-        <add key="OpenRasta-Core CI" value="https://ci.appveyor.com/nuget/openrasta-core" />
-    </packageSources>
-</configuration>
