+++ src/OpenRasta.DI.Windsor/OpenRasta.DI.Windsor.csproj
@@ -18,6 +18,8 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Castle.Windsor" Version="4.1.0" />
-    <PackageReference Include="openrasta-core" Version="2.6.0-preview.1.1328" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\OpenRasta\OpenRasta.csproj" />
   </ItemGroup>
 </Project>
