--- src/openrasta-castle-windsor.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2010
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenRasta.DI.Windsor", "OpenRasta.DI.Windsor\OpenRasta.DI.Windsor.csproj", "{ED7BA1F8-6257-4BDC-B6BE-F22D9C5BC140}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenRasta.DI.Windsor.Tests.Unit", "OpenRasta.DI.Windsor.Tests.Unit\OpenRasta.DI.Windsor.Tests.Unit.csproj", "{9070DF13-BC24-4935-BAC5-49A4876960E9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ED7BA1F8-6257-4BDC-B6BE-F22D9C5BC140}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ED7BA1F8-6257-4BDC-B6BE-F22D9C5BC140}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ED7BA1F8-6257-4BDC-B6BE-F22D9C5BC140}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ED7BA1F8-6257-4BDC-B6BE-F22D9C5BC140}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9070DF13-BC24-4935-BAC5-49A4876960E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9070DF13-BC24-4935-BAC5-49A4876960E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9070DF13-BC24-4935-BAC5-49A4876960E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9070DF13-BC24-4935-BAC5-49A4876960E9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F05EF29D-75AD-436B-A09F-7198F05C9CDE}
-	EndGlobalSection
-EndGlobal
